Remove protobuf and change on existing implementation + PR feedback.
