+++ next-gen/src/OpenTelemetry.AutoInstrumentation.Sdk/OpenTelemetryProtocol/OtlpExporterFactory.cs
@@ -68,8 +68,8 @@ public sealed class OtlpExporterFactory
     {
         ArgumentNullException.ThrowIfNull(options);
 
-        return new ProtobufOtlpSpanExporterAsync(
-            _LoggerFactory.CreateLogger<ProtobufOtlpSpanExporterAsync>(),
+        return new OtlpSpanExporterAsync(
+            _LoggerFactory.CreateLogger<OtlpSpanExporterAsync>(),
             options);
     }
 }
